--- src/adapters/index.ts
@@ -1,1 +1,0 @@
-export { withRetry, type RetryOptions } from './retry'
