+++ src/index.spec.ts
@@ -95,6 +95,7 @@ describe('reject', () => {
 
       const upfetch = up(fetch, (input) => ({
          baseUrl: 'https://example.com',
+         retry: { attempts: 0 },
       }))
 
       await upfetch('').catch((error) => {
@@ -115,6 +116,7 @@ describe('reject', () => {
 
       const upfetch = up(fetch, () => ({
          baseUrl: 'https://example.com',
+         retry: { attempts: 0 },
          reject: () => false,
       }))
 
@@ -135,6 +137,7 @@ describe('reject', () => {
 
       const upfetch = up(fetch, () => ({
          baseUrl: 'https://example.com',
+         retry: { attempts: 0 },
          reject: () => {
             expect(catchCount).toBe(0)
             catchCount++
@@ -165,6 +168,7 @@ describe('reject', () => {
 
       const upfetch = up(fetch, () => ({
          baseUrl: 'https://example.com',
+         retry: { attempts: 0 },
          reject: () => {
             expect(catchCount).toBe(0)
             catchCount++
@@ -196,6 +200,7 @@ describe('reject', () => {
 
       const upfetch = up(fetch, () => ({
          baseUrl: 'https://example.com',
+         retry: { attempts: 0 },
          reject: async () => {
             return new Promise((resolve) => {
                catchCount++
@@ -737,6 +742,7 @@ describe('parseRejected', () => {
 
       const upfetch = up(fetch, () => ({
          baseUrl: 'https://example.com',
+         retry: { attempts: 0 },
       }))
       await upfetch('').catch((error) => {
          expect(error.data).toEqual({ hello: 'world' })
@@ -752,6 +758,7 @@ describe('parseRejected', () => {
 
       const upfetch = up(fetch, () => ({
          baseUrl: 'https://example.com',
+         retry: { attempts: 0 },
       }))
       await upfetch('').catch((error) => {
          expect(error.data).toEqual('some text')
@@ -767,6 +774,7 @@ describe('parseRejected', () => {
 
       const upfetch = up(fetch, () => ({
          baseUrl: 'https://example.com',
+         retry: { attempts: 0 },
          parseRejected(res, request) {
             expect(res instanceof Response).toEqual(true)
             expect(request.url).toEqual('https://example.com/')
@@ -789,6 +797,7 @@ describe('parseRejected', () => {
 
       const upfetch = up(fetch, () => ({
          baseUrl: 'https://example.com',
+         retry: { attempts: 0 },
          parseRejected(res) {
             expect(count).toEqual(1)
             count++
@@ -815,6 +824,7 @@ describe('parseRejected', () => {
 
       const upfetch = up(fetch, () => ({
          baseUrl: 'https://example.com',
+         retry: { attempts: 0 },
          parseRejected(res) {
             expect(res.body).toEqual(null)
             expect(count).toEqual(1)
@@ -835,6 +845,7 @@ describe('parseRejected', () => {
 
       const upfetch = up(fetch, () => ({
          baseUrl: 'https://example.com',
+         retry: { attempts: 0 },
          parseRejected: () => Promise.resolve('from=up'),
       }))
       await upfetch('', {
@@ -888,6 +899,7 @@ describe('onError', () => {
 
       const upfetch = up(fetch, () => ({
          baseUrl: 'https://example.com',
+         retry: { attempts: 0 },
          onError(error, request) {
             if (isResponseError(error)) {
                exec++
@@ -1225,4 +1237,452 @@ describe('timeout', () => {
          })
       })
    }
+})
+describe('retry', () => {
+   const baseUrl = 'https://example.com'
+
+   test('should call retry.when with the response', async () => {
+      const upfetch = up(fetch, () => ({
+         baseUrl,
+         reject: () => false,
+         retry: {
+            attempts: 1,
+            delay: 100,
+            when(ctx) {
+               spy()
+               if (ctx.response) {
+                  expect(ctx.response instanceof Response).toBe(true)
+               } else {
+                  expect(ctx.error instanceof Error).toBe(true)
+               }
+               return false
+            },
+         },
+      }))
+      server.use(
+         http.get(baseUrl, () =>
+            HttpResponse.json({ hello: 'world' }, { status: 500 }),
+         ),
+      )
+      const spy = vi.fn()
+
+      await upfetch('/')
+      expect(spy).toHaveBeenCalledTimes(1)
+   })
+
+   test('should not call retry.when or retry.delay when retry.attempts returns 0', async () => {
+      const upfetch = up(fetch, () => ({
+         baseUrl,
+         reject: () => false,
+         retry: {
+            when: whenSpy,
+            attempts: () => 0,
+            delay: () => {
+               delaySpy()
+               return 0
+            },
+         },
+      }))
+      server.use(
+         http.get(baseUrl, () =>
+            HttpResponse.json({ hello: 'world' }, { status: 500 }),
+         ),
+      )
+      const whenSpy = vi.fn()
+      const delaySpy = vi.fn()
+
+      await upfetch('/')
+      expect(whenSpy).not.toHaveBeenCalled()
+      expect(delaySpy).not.toHaveBeenCalled()
+   })
+
+   test('should call retry.attempts with the request when retry.when returns true', async () => {
+      const upfetch = up(fetch, () => ({
+         baseUrl,
+         reject: () => false,
+         retry: {
+            when: () => true,
+            delay: 100,
+            attempts({ request }) {
+               spy()
+               expectTypeOf(request).toEqualTypeOf<Request>()
+               expect(request instanceof Request).toBe(true)
+               return 0
+            },
+         },
+      }))
+
+      server.use(
+         http.get(baseUrl, () =>
+            HttpResponse.json({ hello: 'world' }, { status: 500 }),
+         ),
+      )
+      const spy = vi.fn()
+
+      await upfetch('/')
+      expect(spy).toHaveBeenCalledTimes(1)
+   })
+
+   test('should not call retry.delay when retry.attempts returns 0', async () => {
+      const upfetch = up(fetch, () => ({
+         baseUrl,
+         reject: () => false,
+         retry: {
+            when: () => true,
+            attempts: () => 0,
+            delay: () => {
+               spy()
+               return 0
+            },
+         },
+      }))
+      server.use(
+         http.get(baseUrl, () =>
+            HttpResponse.json({ hello: 'world' }, { status: 500 }),
+         ),
+      )
+      const spy = vi.fn()
+
+      await upfetch('/')
+      expect(spy).toHaveBeenCalledTimes(0)
+   })
+
+   test('should call retry.delay with the attempt number and response when retry.when returns true and retry.attempts returns more than 0', async () => {
+      const upfetch = up(fetch, () => ({
+         baseUrl,
+         reject: () => false,
+         retry: {
+            when: () => true,
+            attempts: () => 1,
+            delay({ attempt, response, error }) {
+               spy()
+               expectTypeOf(attempt).toEqualTypeOf<number>()
+               expect(attempt).toBe(1)
+               if (response) {
+                  expectTypeOf(response).toEqualTypeOf<Response>()
+               } else {
+                  expectTypeOf(error).toEqualTypeOf<{}>()
+               }
+               expect(response instanceof Response).toBe(true)
+               return 0
+            },
+         },
+      }))
+      server.use(
+         http.get(baseUrl, () =>
+            HttpResponse.json({ hello: 'world' }, { status: 500 }),
+         ),
+      )
+      const spy = vi.fn()
+
+      await upfetch('/')
+      expect(spy).toHaveBeenCalledTimes(1)
+   })
+
+   test('should retry `N = retry.attempts()` times when retry.when returns true', async () => {
+      const upfetch = up(fetch, () => ({
+         baseUrl,
+         reject: () => false,
+         retry: {
+            attempts: 1,
+            delay: 100,
+            when: () => true,
+         },
+      }))
+      const spy = vi.fn()
+
+      server.use(
+         http.get(baseUrl, async () => {
+            spy()
+            return HttpResponse.json({ hello: 'world' }, { status: 500 })
+         }),
+      )
+
+      await upfetch('/')
+      expect(spy).toHaveBeenCalledTimes(2)
+   })
+
+   test('should call retry.attempts, then retry.when then retry.delay', async () => {
+      let exec = 0
+      const upfetch = up(fetch, () => ({
+         baseUrl,
+         reject: () => false,
+         retry: {
+            attempts: () => {
+               expect(++exec).toBe(1)
+               return 1
+            },
+            when: () => {
+               expect(++exec).toBe(2)
+               return true
+            },
+            delay: () => {
+               expect(++exec).toBe(3)
+               return 0
+            },
+         },
+      }))
+
+      server.use(
+         http.get(baseUrl, async () => {
+            return HttpResponse.json({ hello: 'world' }, { status: 500 })
+         }),
+      )
+
+      await upfetch('/', {
+         retry: { when: () => false },
+      })
+   })
+
+   test('should allow upfetch.attempts to override up.attempts', async () => {
+      const upfetch = up(fetch, () => ({
+         baseUrl,
+         reject: () => false,
+         retry: {
+            when: () => true,
+            delay: 100,
+            attempts: 1,
+         },
+      }))
+      const spy = vi.fn()
+
+      server.use(
+         http.get(baseUrl, async () => {
+            spy()
+            return HttpResponse.json({ hello: 'world' }, { status: 500 })
+         }),
+      )
+
+      await upfetch('/', {
+         retry: { attempts: 2 },
+      })
+      // no retry
+      expect(spy).toHaveBeenCalledTimes(3)
+   })
+
+   test('should handle errors during attempts() function execution', async () => {
+      const upfetch = up(fetch, () => ({
+         baseUrl,
+         reject: () => false,
+         retry: {
+            when: () => true,
+            delay: 100,
+            attempts: () => {
+               throw new Error('attempts error')
+            },
+         },
+      }))
+
+      server.use(
+         http.get(baseUrl, async () => HttpResponse.json({}, { status: 500 })),
+      )
+
+      let exec = 0
+
+      await upfetch('/').catch((error) => {
+         exec++
+         expect(error.name).toBe('Error')
+      })
+      expect(exec).toBe(1)
+   })
+
+   test('should handle errors during retry.delay function execution', async () => {
+      const upfetch = up(fetch, () => ({
+         baseUrl,
+         reject: () => false,
+         retry: {
+            when: () => true,
+            attempts: 1,
+            delay: () => {
+               throw new Error('delay error')
+            },
+         },
+      }))
+      const spy = vi.fn()
+
+      server.use(
+         http.get(baseUrl, async () => {
+            spy()
+            return HttpResponse.json({ hello: 'world' }, { status: 500 })
+         }),
+      )
+
+      await expect(upfetch('/')).rejects.toThrow('delay error')
+      expect(spy).toHaveBeenCalledTimes(1) // Only initial request, no retries due to error
+   })
+
+   test('should not retry when request times out', async () => {
+      const upfetch = up(fetch, () => ({
+         baseUrl,
+         timeout: 100, // Set a short timeout
+         reject: () => false,
+         retry: {
+            when: () => true,
+            delay: 100,
+            attempts: 2,
+         },
+      }))
+
+      server.use(
+         http.get(baseUrl, async () => {
+            await new Promise((resolve) => setTimeout(resolve, 200)) // Delay longer than timeout
+            return HttpResponse.json({ hello: 'world' }, { status: 500 })
+         }),
+      )
+
+      let exec = 0
+
+      await upfetch('/').catch((error) => {
+         exec++
+         expect(error.name).toBe('TimeoutError')
+      })
+      expect(exec).toBe(1)
+   })
+
+   test('should not retry when request is aborted', async () => {
+      const controller = new AbortController()
+      const upfetch = up(fetch, () => ({
+         baseUrl,
+         reject: () => false,
+         retry: {
+            when: () => true,
+            delay: 100,
+            attempts: 2,
+         },
+      }))
+
+      server.use(
+         http.get(baseUrl, async () => {
+            await new Promise((resolve) => setTimeout(resolve, 100)) // Add delay to ensure we can abort
+            return HttpResponse.json({ hello: 'world' }, { status: 500 })
+         }),
+      )
+
+      let exec = 0
+
+      const promise = upfetch('/', { signal: controller.signal })
+      controller.abort()
+      await promise.catch((error) => {
+         exec++
+         expect(error.name).toBe('AbortError')
+      })
+      expect(exec).toBe(1)
+   })
+
+   test('should not call onRetry when no retry is needed', async () => {
+      const onRetrySpy = vi.fn()
+      const upfetch = up(fetch, () => ({
+         baseUrl,
+         onRetry: onRetrySpy,
+         reject: () => false,
+         retry: {
+            when: () => false,
+            attempts: 2,
+            delay: 0,
+         },
+      }))
+
+      server.use(
+         http.get(baseUrl, () => HttpResponse.json({}, { status: 500 })),
+      )
+
+      await upfetch('/')
+      expect(onRetrySpy).not.toHaveBeenCalled()
+   })
+
+   test('should call onRetry before each retry attempt', async () => {
+      const onRetrySpy = vi.fn()
+      const upfetch = up(fetch, () => ({
+         baseUrl,
+         onRetry: onRetrySpy,
+         reject: () => false,
+         retry: {
+            when: () => true,
+            attempts: 2,
+            delay: 0,
+         },
+      }))
+
+      server.use(
+         http.get(baseUrl, () => HttpResponse.json({}, { status: 500 })),
+      )
+
+      await upfetch('/')
+      expect(onRetrySpy).toHaveBeenCalledTimes(2)
+      expect(onRetrySpy).toHaveBeenNthCalledWith(1, {
+         attempt: 1,
+         response: expect.any(HttpResponse),
+         request: expect.any(Request),
+      })
+      expect(onRetrySpy).toHaveBeenNthCalledWith(2, {
+         attempt: 2,
+         response: expect.any(HttpResponse),
+         request: expect.any(Request),
+      })
+   })
+
+   test('should execute both up.onRetry and upfetch.onRetry', async () => {
+      const defaultSpy = vi.fn()
+      const fetcherSpy = vi.fn()
+
+      let exec = 0
+
+      const upfetch = up(fetch, () => ({
+         baseUrl,
+         onRetry(context) {
+            expect(++exec).toBe(1)
+            defaultSpy(context)
+         },
+         reject: () => false,
+         retry: {
+            when: () => true,
+            attempts: 1,
+            delay: 0,
+         },
+      }))
+
+      server.use(
+         http.get(baseUrl, () => HttpResponse.json({}, { status: 500 })),
+      )
+
+      await upfetch('/', {
+         onRetry(context) {
+            expect(++exec).toBe(2)
+            fetcherSpy(context)
+         },
+      })
+      expect(defaultSpy).toHaveBeenCalledTimes(1)
+      expect(fetcherSpy).toHaveBeenCalledTimes(1)
+   })
+
+   test('should abort retry immediately if signal controller aborts during retry delay', async () => {
+      const controller = new AbortController()
+      const upfetch = up(fetch, () => ({
+         baseUrl,
+         reject: () => false,
+         retry: {
+            when: () => true,
+            attempts: 2,
+            delay: 1000,
+         },
+      }))
+
+      server.use(
+         http.get(baseUrl, async () => HttpResponse.json({}, { status: 500 })),
+      )
+
+      let exec = 0
+
+      const promise = upfetch('/', { signal: controller.signal })
+      await scheduler.wait(50)
+      const now = Date.now()
+      controller.abort()
+      await promise.catch((error) => {
+         exec++
+         expect(error.name).toBe('AbortError')
+      })
+      expect(exec).toBe(1)
+      // should not wait for the whole 1000ms delay
+      expect(Date.now() - now).toBeLessThan(50)
+   })
 })
