+++ tsup.config.ts
@@ -1,10 +1,7 @@
 import { defineConfig } from 'tsup'
 
 export default defineConfig({
-   entry: {
-      index: 'src/index.ts',
-      'adapters': 'src/adapters/index.ts',
-   },
+   entry: { index: 'src/index.ts' },
    sourcemap: true,
    clean: true,
    // fix pnpm workspaces module resolution issues
